Make agent instances inherit a _COPY_ of the set of understood actions from it's parent
This is important, without the #dup call, you get _the actual set_ of actions from your parent, which means you destructively update your parents understanding of actions. This is bad.

There is a better way to structure this, I'm sure. Perhaps a non-private "understood_actions" method which returns a dup? Would that be confusing?